% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_correlation.R
\name{bayesian_correlation}
\alias{bayesian_correlation}
\title{Compute Bayesian Correlation with Optional Plot if requested}
\usage{
bayesian_correlation(data, x, y, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{x}{The name of the first continuous variable.}

\item{y}{The name of the second continuous variable.}

\item{plot}{A logical value indicating whether to generate a scatter plot with the correlation line (default is TRUE).}
}
\value{
An object of class \code{BFBayesFactor} containing the results of the Bayesian correlation analysis.
}
\description{
This function computes the Bayesian correlation between two continuous variables using the BayesFactor package.
It also generates a scatter plot with the correlation line if requested.
}
\examples{
bayesian_correlation(mtcars, "mpg", "wt")
}
