% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_lm.R
\name{bayesian_lm}
\alias{bayesian_lm}
\title{Fit a Bayesian Linear Regression Model and a plot if requested}
\usage{
bayesian_lm(data, formula, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{formula}{A formula specifying the model.}

\item{plot}{A logical value indicating whether to generate a plot of the posterior distribution (default is TRUE).}
}
\value{
An object of class \code{BFBayesFactor} containing the results of the Bayesian linear regression.
}
\description{
This function fits a Bayesian linear regression model and returns the results.
It also generates a plot of the posterior distribution if requested.
}
\examples{
bayesian_lm(mtcars, mpg ~ wt)
}
