% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_t_test.R
\name{bayesian_t_test}
\alias{bayesian_t_test}
\title{Perform Bayesian T-Test with Optional Plot if requested}
\usage{
bayesian_t_test(data, group_col, value_col, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group_col}{The name of the column in the data frame that defines the groups (factor or character).}

\item{value_col}{The name of the numeric column in the data frame for which the t-test is to be performed.}

\item{plot}{A logical value indicating whether to generate a density plot of the posterior distribution (default is TRUE).}
}
\value{
An object of class \code{BFBayesFactor} containing the results of the Bayesian t-test.
}
\description{
This function performs a Bayesian t-test to compare means between two groups using the BayesFactor package.
It also generates a density plot of the posterior distribution if requested.
}
\examples{
bayesian_t_test(mtcars, "am", "mpg")
}
