% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_test.R
\name{chi_square_test}
\alias{chi_square_test}
\title{Perform Chi-Square Test with a optional Plot}
\usage{
chi_square_test(data, col1, col2, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{col1}{The name of the first categorical column in the data frame.}

\item{col2}{The name of the second categorical column in the data frame.}

\item{plot}{A logical value indicating whether to generate a mosaic plot (default is TRUE).}
}
\value{
An object of class \code{htest} containing the results of the chi-square test.
}
\description{
This function performs a chi-square test of independence for two categorical variables and optionally generates a mosaic plot for visualazation.
}
\examples{
chi_square_test(mtcars, "cyl", "gear")
}
