% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_lm.R
\name{simple_lm}
\alias{simple_lm}
\title{Fit a Simple Linear Regression Model with a optional plot (if requested)}
\usage{
simple_lm(data, formula, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{formula}{A formula specifying the model.}

\item{plot}{A logical value indicating whether to generate plots (default is TRUE).}
}
\value{
A summary of the linear regression model.
}
\description{
This function fits a simple linear regression model and returns the summary of the model.
It also generates a scatter plot with the regression line and a residual plot if requested.
}
\examples{
simple_lm(mtcars, mpg ~ wt)
}
