% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Statistics}
\usage{
summary_stats(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame of summary statistics for each numeric column.
}
\description{
This function gives back summary statistics (mean, median, standard deviation, Min and Max) for numeric columns in a dataset.
}
\examples{
summary_stats(mtcars)
}
