% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Perform T-Test with a optional integrated plot}
\usage{
t_test(data, group_col, value_col, plot = TRUE)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group_col}{The name of the column in the data frame that defines the groups (factor or character).}

\item{value_col}{The name of the numeric column in the data frame for which the t-test is to be performed.}

\item{plot}{A logical value indicating whether to generate a boxplot (default is TRUE).}
}
\value{
An object of class \code{htest} containing the results of the t-test.
}
\description{
This function performs a t-test to compare means between two groups and optionally generates a boxplot to visualize the distribution.
}
\examples{
t_test(mtcars, "am", "mpg")
}
